+++ Source/ResourceImporter/UserInterface/Private/Windows/AnimationElementsWindow.cpp
@@ -41,7 +41,7 @@ AnimationElementsWindow::AnimationElementsWindow(MainScreen& inScreen,
                                   {1, 1, 1, 1});
 
     auto styleLabel
-        = [&](AUI::Text& label, const std::string& text, int fontSize) {
+        = [&](AUI::Text& label, const std::string& text, float fontSize) {
         label.setFont((Paths::FONT_DIR + "B612-Regular.ttf"), fontSize);
         label.setColor({255, 255, 255, 255});
         label.setText(text);
