+++ Source/ResourceImporter/UserInterface/Private/Windows/BoundingBoxPropertiesWindow.cpp
@@ -66,7 +66,7 @@ BoundingBoxPropertiesWindow::BoundingBoxPropertiesWindow(
                                   {1, 1, 1, 1});
 
     auto styleLabel
-        = [&](AUI::Text& label, const std::string& text, int fontSize) {
+        = [&](AUI::Text& label, const std::string& text, float fontSize) {
         label.setFont((Paths::FONT_DIR + "B612-Regular.ttf"), fontSize);
         label.setColor({255, 255, 255, 255});
         label.setText(text);
